requirements.txt
praw==7.7.1
python-dotenv==1.0.1
groq==0.4.2
requests>=2.31.0
